Extract the dependency tracking to DependencyTracker
